share one group authority read fixture across the mutation tests
The lifecycle and transport mutation tests each built the same
GroupMutationRead decision surface under different names. The facts and
audit-stamp builders were byte-identical; the read builders differed only
in their option bag, where transport's was the superset.

Move the transport implementation into group-mutation-test-runtime.ts as
createGroupAuthorityRead / createGroupAuthorityFacts /
createGroupAuthorityAuditStamp and let both files call it. The lifecycle
file's connectRead and criterionRead now spread the shared builder, and
its three policyStatus call sites re-key onto the superset's policy
option ('present' had no caller and folds into 'optimistic').

No assertion or test title changes.
