+++ packages/tests/shared-server/rallar-system/group-state/mutation/group-transport-mutation.test.ts
@@ -1,32 +1,26 @@
 import { describe, expect, it } from 'vitest';
 
-import type {
-    GroupMutationCommand,
-    GroupMutationFacts,
-    GroupMutationRead
-} from '@shared-server/rallar-system/group-state/mutation/group-mutation-contracts.ts';
+import type { GroupMutationCommand } from '@shared-server/rallar-system/group-state/mutation/group-mutation-contracts.ts';
 import { GROUP_MUTATION_INTERNAL_AUTHORITY_MODES } from '@shared-server/rallar-system/group-state/mutation/group-mutation-contracts.ts';
 import { computeGroupMutation } from '@shared-server/rallar-system/group-state/mutation/orchestration/compute-group-mutation.ts';
 import { GroupPolicyDeniedError } from '@shared-server/rallar-system/group-state/policy/group-policy-result.ts';
-import { resolveGroupLifecyclePolicyPreset } from '@shared/api/group-lifecycle/group-lifecycle-policy-presets.ts';
 import { GROUP_LIFECYCLE_STATES } from '@shared/api/group-lifecycle/group-lifecycle-policy.ts';
 import type { GroupPolicyDenied } from '@shared/api/group-policy-types.ts';
-import type { AuditStamp, Group } from '@shared/api/group-types.ts';
+import type { Group } from '@shared/api/group-types.ts';
 import { GROUP_PRESENCE_SUMMARY_TOPIC } from '@shared/queuebox/GroupPresenceSummaryEntryContract.ts';
-import { createTestGroup } from '../../../../create-test-group.ts';
 
-import { groupMemberStorageKey, groupRef, groupStorageKey, storedEntry } from './group-mutation-test-runtime.ts';
+import { createGroupAuthorityAuditStamp, createGroupAuthorityFacts, createGroupAuthorityRead, groupRef } from './group-mutation-test-runtime.ts';
 
 describe('group transport mutation computation', () => {
     it('pauses a flowing group by halting transport alone', () => {
         const computed = computeGroupMutation({
             command: transportCommand('pauseGroupTransport'),
-            read: transportRead({
+            read: createGroupAuthorityRead({
                 transportState: 'flowing',
                 lifecycleState: 'active',
                 formationEpoch: 3
             }),
-            facts: transportFacts()
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('write');
@@ -46,8 +40,8 @@ describe('group transport mutation computation', () => {
     it('resumes a halted group by restoring flow', () => {
         const computed = computeGroupMutation({
             command: transportCommand('resumeGroupTransport'),
-            read: transportRead({ transportState: 'halted', lifecycleState: 'reconnecting' }),
-            facts: transportFacts()
+            read: createGroupAuthorityRead({ transportState: 'halted', lifecycleState: 'reconnecting' }),
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('write');
@@ -64,8 +58,8 @@ describe('group transport mutation computation', () => {
     ])('answers no-op when %s finds the valve already at %s', (operation, transportState) => {
         const computed = computeGroupMutation({
             command: transportCommand(operation),
-            read: transportRead({ transportState }),
-            facts: transportFacts()
+            read: createGroupAuthorityRead({ transportState }),
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('no-op');
@@ -81,12 +75,12 @@ describe('group transport mutation computation', () => {
     it('arms no formation timer: the only outbox entry is the presence summary', () => {
         const computed = computeGroupMutation({
             command: transportCommand('pauseGroupTransport'),
-            read: transportRead({
+            read: createGroupAuthorityRead({
                 transportState: 'flowing',
                 lifecycleState: 'connecting',
                 establishmentStartedAtEpochMs: 1_500
             }),
-            facts: transportFacts()
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('write');
@@ -109,8 +103,8 @@ describe('group transport mutation computation', () => {
     it.each(GROUP_LIFECYCLE_STATES)('halts a flowing group in the %s stage', (lifecycleState) => {
         const computed = computeGroupMutation({
             command: transportCommand('pauseGroupTransport'),
-            read: transportRead({ transportState: 'flowing', lifecycleState }),
-            facts: transportFacts()
+            read: createGroupAuthorityRead({ transportState: 'flowing', lifecycleState }),
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('write');
@@ -127,8 +121,8 @@ describe('group transport mutation computation', () => {
         const denial = readTransportDenial(() =>
             computeGroupMutation({
                 command: transportCommand('pauseGroupTransport'),
-                read: transportRead({ transportState: 'flowing' }, { policy: 'server-auto' }),
-                facts: transportFacts()
+                read: createGroupAuthorityRead({ transportState: 'flowing' }, { policy: 'server-auto' }),
+                facts: createGroupAuthorityFacts()
             })
         );
 
@@ -143,7 +137,7 @@ describe('group transport mutation computation', () => {
         const denial = readTransportDenial(() =>
             computeGroupMutation({
                 command: transportCommand('resumeGroupTransport', 'bob'),
-                read: transportRead(
+                read: createGroupAuthorityRead(
                     { transportState: 'halted' },
                     {
                         policy: 'managed',
@@ -151,7 +145,7 @@ describe('group transport mutation computation', () => {
                         activeMemberPrincipalIds: ['alice', 'bob']
                     }
                 ),
-                facts: transportFacts('bob')
+                facts: createGroupAuthorityFacts('bob')
             })
         );
 
@@ -162,8 +156,8 @@ describe('group transport mutation computation', () => {
     it('allows the valve to the manager the same policy resolves', () => {
         const computed = computeGroupMutation({
             command: transportCommand('resumeGroupTransport'),
-            read: transportRead({ transportState: 'halted' }, { policy: 'managed' }),
-            facts: transportFacts()
+            read: createGroupAuthorityRead({ transportState: 'halted' }, { policy: 'managed' }),
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('write');
@@ -173,8 +167,8 @@ describe('group transport mutation computation', () => {
         const denial = readTransportDenial(() =>
             computeGroupMutation({
                 command: transportCommand('pauseGroupTransport'),
-                read: transportRead({ transportState: 'flowing' }, { actorIsMember: false }),
-                facts: transportFacts()
+                read: createGroupAuthorityRead({ transportState: 'flowing' }, { actorIsMember: false }),
+                facts: createGroupAuthorityFacts()
             })
         );
 
@@ -189,12 +183,12 @@ describe('group transport mutation computation', () => {
             const denial = readTransportDenial(() =>
                 computeGroupMutation({
                     command: transportCommand('pauseGroupTransport'),
-                    read: transportRead({
+                    read: createGroupAuthorityRead({
                         transportState: 'flowing',
                         status,
-                        [status]: transportAuditStamp(1_500, 'alice')
+                        [status]: createGroupAuthorityAuditStamp(1_500, 'alice')
                     }),
-                    facts: transportFacts()
+                    facts: createGroupAuthorityFacts()
                 })
             );
 
@@ -205,8 +199,8 @@ describe('group transport mutation computation', () => {
     it('fails closed on an unreadable stored policy instead of reading it as permissive', () => {
         const computed = computeGroupMutation({
             command: transportCommand('pauseGroupTransport'),
-            read: transportRead({ transportState: 'flowing' }, { policy: 'corrupt' }),
-            facts: transportFacts()
+            read: createGroupAuthorityRead({ transportState: 'flowing' }, { policy: 'corrupt' }),
+            facts: createGroupAuthorityFacts()
         });
 
         expect(computed.outcome).toBe('rejected');
@@ -225,8 +219,8 @@ describe('group transport mutation computation', () => {
         expect(() =>
             computeGroupMutation({
                 command: internalTransportCommand(),
-                read: transportRead({ transportState: 'flowing' }),
-                facts: { ...transportFacts(), internalAuthority, authenticatedAuthority: null }
+                read: createGroupAuthorityRead({ transportState: 'flowing' }),
+                facts: { ...createGroupAuthorityFacts(), internalAuthority, authenticatedAuthority: null }
             })
         ).toThrowError(TypeError);
     });
@@ -276,107 +270,4 @@ function transportCommand(
             traceId: null
         }
     } as GroupMutationCommand;
-}
-
-interface TransportReadOptions {
-    readonly policy?: 'absent' | 'corrupt' | 'optimistic' | 'managed' | 'server-auto';
-    readonly actorPrincipalId?: string;
-    readonly actorIsMember?: boolean;
-    readonly activeMemberPrincipalIds?: readonly string[];
-}
-
-function transportRead(
-    groupOverrides: Partial<Group>,
-    options: TransportReadOptions = {}
-): GroupMutationRead {
-    const actorPrincipalId = options.actorPrincipalId ?? 'alice';
-    const actorMember = transportActorMember(actorPrincipalId);
-    return {
-        idempotency: null,
-        group: storedEntry(groupStorageKey(), createTestGroup({ ...groupRef('pure-room'), ...groupOverrides })),
-        expiredGroupEntry: null,
-        actorMember: options.actorIsMember === false ? null : actorMember,
-        targetMember: null,
-        authorityMember: null,
-        directorMember: null,
-        actorMemberEntry: options.actorIsMember === false
-            ? null
-            : storedEntry(groupMemberStorageKey(actorPrincipalId), actorMember),
-        targetMemberEntry: null,
-        authorityMemberEntry: null,
-        directorMemberEntry: null,
-        targetPresence: null,
-        expiredTargetPresenceEntry: null,
-        targetAdmission: null,
-        authorityAdmission: null,
-        directorAdmission: null,
-        authorityPresenceSessions: [],
-        authorityPresenceSessionEntries: [],
-        presenceSummary: null,
-        lifecyclePolicy: transportPolicyRead(options.policy ?? 'absent'),
-        activeMemberPrincipalIds: options.activeMemberPrincipalIds ??
-            (options.actorIsMember === false ? [] : [actorPrincipalId]),
-        plannedLayoutRow: null,
-        acceptedLayoutRow: null
-    } as GroupMutationRead;
-}
-
-function transportActorMember(principalId: string) {
-    const audit = transportAuditStamp(1_000, principalId);
-    return {
-        ...groupRef('pure-room'),
-        principalId,
-        role: 'member' as const,
-        status: 'active' as const,
-        invitedByPrincipalId: null,
-        invitationExpiresAtEpochMs: null,
-        left: null,
-        removed: null,
-        banned: null,
-        joined: audit,
-        updated: audit
-    };
-}
-
-function transportPolicyRead(requested: NonNullable<TransportReadOptions['policy']>) {
-    if (requested === 'corrupt') {
-        return { status: 'corrupt' as const, reason: 'stored policy is not an object' };
-    }
-    if (requested === 'absent') {
-        return { status: 'absent' as const };
-    }
-    return {
-        status: 'present' as const,
-        policy: requested === 'server-auto'
-            ? { ...resolveGroupLifecyclePolicyPreset('optimistic'), initiator: 'server-auto' as const }
-            : resolveGroupLifecyclePolicyPreset(requested)
-    };
-}
-
-function transportFacts(principalId = 'alice'): GroupMutationFacts {
-    return {
-        nowEpochMs: 2_000,
-        expireAtEpochMs: 253_402_300_799_999,
-        serviceId: 'group-service',
-        eventId: 'event-1',
-        commandHash: `sha256:${'a'.repeat(64)}`,
-        attemptCount: 1,
-        resolvedJoinCode: null,
-        joinCodeVerifier: null,
-        internalAuthority: 'none',
-        authenticatedAuthority: {
-            principalId,
-            sessionId: `${principalId}-session`
-        }
-    };
-}
-
-function transportAuditStamp(atEpochMs: number, principalId: string): AuditStamp {
-    return {
-        atEpochMs,
-        actor: { kind: 'principal', principalId },
-        reason: null,
-        traceId: null,
-        requestId: 'seed'
-    };
 }
